requirements.txt
# Core
flask>=2.3.0
pymongo>=4.5.0
openai>=1.0.0
python-dotenv>=1.0.0

# Utilities
geopy>=2.3.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Testing (dev-only)
pytest>=7.4.0
requests-mock>=1.11.0
flask-testing>=0.8.0

# Bonus (uncomment if needed)
# redis>=4.6.0
# faker>=19.0.0